requirements.txt
aiogram>=3.4,<4.0
python-dotenv>=1.0,<2.0
ruff>=0.5.0
Pillow>=10.0.0,<12.0.0
pyinstaller>=6.6,<7.0
Flask>=2.0,<4.0
